mediscan-pro/requirements.txt
# ============================================================================
# FILE: requirements.txt
# Production-grade dependencies with version pinning
# ============================================================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# AI/ML Core
torch==2.8.0
torchvision==0.15.2
transformers==4.36.2
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.1.4
scipy==1.11.4

# Computer Vision
opencv-python==4.9.0.80
Pillow==10.2.0
albumentations==1.3.1

# NLP
spacy==3.7.2
nltk==3.8.1
sentence-transformers==2.2.2

# HTTP & Async
httpx==0.26.0
aiohttp==3.9.1
aiofiles==23.2.1

# Caching & Queue
redis==5.0.1
celery==5.3.4

# Database
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
aiosqlite==0.19.0

# Monitoring & Logging
prometheus-client==0.19.0
loguru==0.7.2
sentry-sdk==1.39.2

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
cryptography==41.0.7

# Data Processing
openpyxl==3.1.2
python-dateutil==2.8.2
pytz==2023.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0
faker==22.0.0

# Utilities
tqdm==4.66.1
click==8.1.7
rich==13.7.0